backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pandas==2.2.3
python-dotenv==1.0.1
google-generativeai==0.8.3
pytest==8.3.3
httpx==0.27.2
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
apscheduler==3.10.4
python-multipart==0.0.12
